requirements.txt
algoliasearch-django>=3.0,<4.0
django
djangorestframework
pyYAML
requests
django-cors-headers